use confirmation modal + refactor
